refactor(trip-details): Replace TripDetails component with its OTP-UI counterpart.
