+++ lib/components/narrative/default/default-itinerary.js
@@ -3,7 +3,7 @@ import React from 'react'
 import NarrativeItinerary from '../narrative-itinerary'
 import ItinerarySummary from './itinerary-summary'
 import ItineraryDetails from './itinerary-details'
-import TripDetails from '../trip-details'
+import TripDetails from '../connected-trip-details'
 import TripTools from '../trip-tools'
 import { formatDuration, formatTime } from '../../../util/time'
