+++ lib/components/narrative/line-itin/itin-body.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import isEqual from 'lodash.isequal'
 
-import TripDetails from '../trip-details'
+import TripDetails from '../connected-trip-details'
 import TripTools from '../trip-tools'
 
 import PlaceRow from './place-row'
