+++ lib/components/narrative/printable/printable-itinerary.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 
-import TripDetails from '../trip-details'
+import TripDetails from '../connected-trip-details'
 import { distanceString } from '../../../util/distance'
 import { formatTime, formatDuration } from '../../../util/time'
 import {
