+++ lib/index.js
@@ -34,7 +34,7 @@ import NarrativeRoutingResults from './components/narrative/narrative-routing-re
 import RealtimeAnnotation from './components/narrative/realtime-annotation'
 import SimpleRealtimeAnnotation from './components/narrative/simple-realtime-annotation'
 import TransportationNetworkCompanyLeg from './components/narrative/default/tnc-leg'
-import TripDetails from './components/narrative/trip-details'
+import TripDetails from './components/narrative/connected-trip-details'
 import TripTools from './components/narrative/trip-tools'
 import LineItinerary from './components/narrative/line-itin/line-itinerary'
