+++ package.json
@@ -29,9 +29,10 @@
   "dependencies": {
     "@conveyal/lonlat": "^1.1.0",
     "@mapbox/polyline": "^0.2.0",
-    "@opentripplanner/from-to-location-picker": "^0.0.14",
-    "@opentripplanner/geocoder": "^0.0.16",
-    "@opentripplanner/location-field": "^0.0.16",
+    "@opentripplanner/from-to-location-picker": "^0.0.18",
+    "@opentripplanner/geocoder": "^0.0.18",
+    "@opentripplanner/location-field": "^0.0.18",
+    "@opentripplanner/trip-details": "^0.0.18",
     "bootstrap": "^3.3.7",
     "clone": "^2.1.0",
     "connected-react-router": "^6.5.2",
